Expose stepper pins through a shared memory location
The work with the Klipper frontend exposed a flaw with how object
commands are implemented.

Klipper requires tight timing when performing motor moves and it
controller when the steps are made using timers that trigger at
specific clock ticks.

Using the established method for submitting object commands had two
major flaws:
   1. It is hard to reliably execute a single command at a time,
      which causes command execution errors and missed steps.
   2. The command execution path was way too slow to support the
      required timing.

This meant that steppers could not be moved successfully with the
Klipper frontend.

To fix this, expose a memory location that serves as stepper pin
control. The stepper object monitors the value in the memory
location and performs actions based on psuedo-pin values. This
enables the stepper object to update it's state much faster and,
in turn, allow Klipper to successfully move the motor.

+++ src/core/common_defs.h
@@ -28,7 +28,7 @@
 typedef core_object_t *(*object_lookup_cb_t)(const core_object_type_t,
                                              const char *, void *);
 typedef core_object_t **(*object_list_cb_t)(const core_object_type_t, void *);
-typedef void (*complete_cb_t)(uint64_t, int, void *);
+typedef void (*complete_cb_t)(uint64_t, int64_t, void *);
 typedef uint64_t (*cmd_submit_cb_t)(core_object_t *, core_object_id_t, uint16_t,
 				    void *, complete_cb_t, void *);
 typedef void (*log_cb_t)(core_log_level_t, core_object_type_t, const char *,
